:hammer: chore: minor refactor
